#8 chore: Core 모듈 간소화
- CoreDTO, CoreTarget, CoreComponent, CoreDataMapper 모듈 -> CoreNetworkService 모듈의 하위 디렉토리로 변경
- CoreError 모듈 -> 각각 CoreNetworkService과 CoreStorageService 모듈의 하위 디렉토리로 변경
