--- Projects/Core/CoreComponent/Project.swift
@@ -1,17 +1,0 @@
-//
-//  Project.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 김상혁 on 2023/04/26.
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project.makeModule(
-    name: Modules.Core(.CoreComponent).name,
-    product: .framework,
-    packages: [],
-    dependencies: [],
-    hasTests: false
-)
