--- Projects/Core/CoreComponent/Sources/EmptyBody.swift
@@ -1,13 +1,0 @@
-//
-//  EmptyBody.swift
-//  CoreComponent
-//
-//  Created by 김상혁 on 2023/04/27.
-//  Copyright © 2023 LifePoop. All rights reserved.
-//
-
-import Foundation
-
-public struct EmptyBody: Encodable {
-    public init() { }
-}
