--- Projects/Core/CoreDIContainer/Sources/CoreDIConatinerSource.swift
@@ -1,8 +1,0 @@
-//
-//  CoreDIConatinerSource.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 김상혁 on 2023/04/29.
-//
-
-import Foundation
