--- Projects/Core/CoreDTO/Project.swift
@@ -1,23 +1,0 @@
-//
-//  Project.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 김상혁 on 2023/04/26.
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project.makeModule(
-    name: Modules.Core(.CoreDTO).name,
-    product: .framework,
-    packages: [
-        .SPM.RxSwift.package
-    ],
-    dependencies: [
-        .SPM.RxSwift.dependency,
-        .SPM.RxCocoa.dependency,
-        .SPM.RxRelay.dependency
-    ],
-    hasTests: false
-)
