--- Projects/Core/CoreDataMapper/Project.swift
@@ -1,25 +1,0 @@
-//
-//  Project.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 김상혁 on 2023/04/29.
-//
-
-import ProjectDescription
-import ProjectDescriptionHelpers
-
-let project = Project.makeModule(
-    name: Modules.Core(.CoreDataMapper).name,
-    product: .framework,
-    packages: [
-        .SPM.RxSwift.package
-    ],
-    dependencies: [
-        .Project.module(.Core(.CoreDTO)).dependency,
-        .Project.module(.Core(.CoreEntity)).dependency,
-        .SPM.RxSwift.dependency,
-        .SPM.RxCocoa.dependency,
-        .SPM.RxRelay.dependency
-    ],
-    hasTests: false
-)
