--- Projects/Core/CoreDataMapper/Sources/Interface/DataMapper.swift
@@ -1,16 +1,0 @@
-//
-//  DataMapper.swift
-//  CoreDataMapper
-//
-//  Created by 김상혁 on 2023/04/27.
-//  Copyright © 2023 LifePoop. All rights reserved.
-//
-
-import Foundation
-
-public protocol DataMapper {
-    associatedtype Input
-    associatedtype Output
-
-    func transform(_ input: Input) -> Output
-}
