+++ Projects/Core/CoreNetworkService/Sources/DTO/CoreExampleDTO.swift
@@ -1,0 +1,19 @@
+//
+//  CoreExampleDTO.swift
+//  CoreNetworkService
+//
+//  Created by 김상혁 on 2023/04/27.
+//  Copyright © 2023 LifePoop. All rights reserved.
+//
+
+import Foundation
+
+public struct CoreExampleDTO: Decodable {
+    public let name: String
+    public let age: Int
+    
+    public init(name: String, age: Int) {
+        self.name = name
+        self.age = age
+    }
+}
