+++ Projects/Core/CoreNetworkService/Sources/HTTPComponent/EmptyBody.swift
@@ -1,0 +1,13 @@
+//
+//  EmptyBody.swift
+//  CoreNetworkService
+//
+//  Created by 김상혁 on 2023/04/27.
+//  Copyright © 2023 LifePoop. All rights reserved.
+//
+
+import Foundation
+
+public struct EmptyBody: Encodable {
+    public init() { }
+}
