--- Projects/Core/CoreTarget/Sources/Interface/TargetType.swift
@@ -1,19 +1,0 @@
-//
-//  TargetType.swift
-//  CoreTarget
-//
-//  Created by 김상혁 on 2023/04/27.
-//  Copyright © 2023 LifePoop. All rights reserved.
-//
-
-import Foundation
-
-import CoreComponent
-
-public protocol TargetType {
-    var baseURL: URL? { get }
-    var path: String { get }
-    var method: HTTPMethod { get }
-    var headers: [String: String]? { get }
-    var parameters: [String: Any]? { get }
-}
