+++ Projects/Features/FeatureLogin/FeatureLoginRepository/Project.swift
@@ -15,13 +15,9 @@ let project = Project.makeModule(
         .SPM.RxSwift.package
     ],
     dependencies: [
-        .Project.module(.Core(.CoreDataMapper)).dependency,
         .Project.module(.Core(.CoreDIContainer)).dependency,
-        .Project.module(.Core(.CoreDTO)).dependency,
         .Project.module(.Core(.CoreEntity)).dependency,
-        .Project.module(.Core(.CoreExtension)).dependency,
         .Project.module(.Core(.CoreNetworkService)).dependency,
-        .Project.module(.Core(.CoreTarget)).dependency,
         .Project.module(.Features(.Login, .UseCase)).dependency,
         .Project.module(.Utils).dependency,
         .SPM.RxSwift.dependency
