+++ Projects/Shared/SharedRepository/Project.swift
@@ -15,15 +15,10 @@ let project = Project.makeModule(
         .SPM.RxSwift.package
     ],
     dependencies: [
-        .Project.module(.Core(.CoreDataMapper)).dependency,
         .Project.module(.Core(.CoreDIContainer)).dependency,
-        .Project.module(.Core(.CoreDTO)).dependency,
         .Project.module(.Core(.CoreEntity)).dependency,
-        .Project.module(.Core(.CoreError)).dependency,
-        .Project.module(.Core(.CoreExtension)).dependency,
         .Project.module(.Core(.CoreNetworkService)).dependency,
         .Project.module(.Core(.CoreStorageService)).dependency,
-        .Project.module(.Core(.CoreTarget)).dependency,
         .Project.module(.Utils).dependency,
         .SPM.RxSwift.dependency
     ]
