+++ Tuist/ProjectDescriptionHelpers/Modules.swift
@@ -75,14 +75,8 @@ public enum FeatureLayerModuleType: CaseIterable {
 }
 
 public enum CoreModuleType: String, CaseIterable {
-    case CoreComponent
-    case CoreDataMapper
     case CoreDIContainer
-    case CoreDTO
     case CoreEntity
-    case CoreError
-    case CoreExtension
     case CoreNetworkService
     case CoreStorageService
-    case CoreTarget
 }
